fix tests from moving stuff around
