+++ tests/unit/alliance/test_alliance_publication.py
@@ -8,7 +8,7 @@ def source_name():
 
 @pytest.fixture
 def script():
-    return "./monarch_ingest/alliance/publication.py"
+    return "./monarch_ingest/ingests/alliance/publication.py"
 
 
 @pytest.fixture
