+++ tests/unit/flybase/test_flybase_publication_to_gene.py
@@ -10,7 +10,7 @@ def source_name():
 # This is the location of the transform code
 @pytest.fixture
 def script():
-    return "./monarch_ingest/flybase/publication_to_gene.py"
+    return "./monarch_ingest/ingests/flybase/publication_to_gene.py"
 
 
 # Create a fixture for a full row, it should be relatively representative of the rows ingested, and can
