+++ tests/unit/pombase/test_pombase_gene_to_phenotype.py
@@ -38,7 +38,7 @@ def entities(mock_koza, global_table):
     return mock_koza(
         name="pombase_gene_to_phenotype",
         data=row,
-        transform_code="./monarch_ingest/pombase/gene_to_phenotype.py",
+        transform_code="./monarch_ingest/ingests/pombase/gene_to_phenotype.py",
         global_table=global_table,
     )
