+++ tests/unit/string/test_string_protein_links.py
@@ -18,7 +18,7 @@ def script():
     """
     :return: string path to STRING protein links ingest script
     """
-    return "./monarch_ingest/string/protein_links.py"
+    return "./monarch_ingest/ingests/string/protein_links.py"
 
 
 @pytest.fixture
