+++ tests/unit/xenbase/test_xenbase_gene.py
@@ -20,7 +20,7 @@ def gene_information_entities(mock_koza):
         ]
     )
 
-    return mock_koza("xenbase_gene", row, "./monarch_ingest/xenbase/gene.py")
+    return mock_koza("xenbase_gene", row, "./monarch_ingest/ingests/xenbase/gene.py")
 
 
 def test_gene_information_gene(gene_information_entities):
