+++ tests/unit/xenbase/test_xenbase_publication_to_gene.py
@@ -62,7 +62,7 @@ def gene_literature_entities(mock_koza, global_table):
     return mock_koza(
         name="xenbase_publication_to_gene",
         data=row,
-        transform_code="./monarch_ingest/xenbase/publication_to_gene.py",
+        transform_code="./monarch_ingest/ingests/xenbase/publication_to_gene.py",
         map_cache=map_cache,
         global_table=global_table,
     )
